Separate out kopia commands from a single file
