+++ pkg/kopia/cmd/server_add_user.go
@@ -1,0 +1,48 @@
+// Copyright 2022 The Kanister Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cmd
+
+import "github.com/kanisterio/kanister/pkg/logsafe"
+
+// ServerAddUser returns the kopia command adding a new user to the Kopia API Server
+func ServerAddUser(
+	encryptionKey,
+	configFilePath,
+	logDirectory,
+	newUsername,
+	userPassword string,
+) []string {
+	return stringSliceCommand(serverAddUser(
+		encryptionKey,
+		configFilePath,
+		logDirectory,
+		newUsername,
+		userPassword,
+	))
+}
+
+func serverAddUser(
+	encryptionKey,
+	configFilePath,
+	logDirectory,
+	newUsername,
+	userPassword string,
+) logsafe.Cmd {
+	args := kopiaArgs(encryptionKey, configFilePath, logDirectory, false)
+	args = args.AppendLoggable(serverSubCommand, userSubCommand, addSubCommand, newUsername)
+	args = args.AppendRedactedKV(userPasswordFlag, userPassword)
+
+	return args
+}
